+++ src/org/lukasj/idea/torquescript/reference/TSNamespaceReference.kt
@@ -22,7 +22,7 @@ class TSNamespaceReference(element: TSIdentExpressionElementImpl, textRange: Tex
 
     override fun resolve(): PsiElement? {
         val resolveResults = multiResolve(false)
-        return if (resolveResults.size == 1) resolveResults[0].element else null
+        return resolveResults.singleOrNull()?.element
     }
 
     override fun getVariants(): Array<Any> {