+++ src/org/lukasj/idea/torquescript/reference/TSObjectReference.kt
@@ -10,9 +10,10 @@ class TSObjectReference(element: PsiNamedElement, textRange: TextRange) : PsiRef
     PsiPolyVariantReference {
     override fun multiResolve(incompleteCode: Boolean): Array<ResolveResult> {
         val project = myElement.project
-        val objects = ReferenceUtil.findObject(project, rangeInElement.substring(element.text))
+        val objects = project.getService(TSTypeLookupService::class.java).findObject(project, rangeInElement.substring(element.text))
 
-        return objects.map { PsiElementResolveResult(it) }
+        return objects.mapNotNull { it.psiElement }
+            .map { PsiElementResolveResult(it) }
             .toTypedArray()
     }